+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for NeoFS Node
 
 ### Added
 - "storage list" command to neofs-lens (#2852)
+- "morph mint-balance" command to neofs-adm (#2867)
 
 ### Fixed
 - GETRANGE request may fail in certain cases (#2849)